+++ frontend/mock-server/test.json
@@ -36,6 +36,22 @@
       "voteCount": 422
     }
   ],
-  "comments": [{ "id": 1, "body": "some comment", "postId": 1 }],
+  "comments": [
+    {
+      "id": 1,
+      "content": "이거참 멋지군요",
+      "createdAt": "2022-03-20 12:20"
+    },
+    {
+      "id": 2,
+      "content": "이거참 멋지군요",
+      "createdAt": "2022-03-20 12:10"
+    },
+    {
+      "id": 3,
+      "content": "이거참 멋지군요",
+      "createdAt": "2022-03-20 12:00"
+    }
+  ],
   "profile": { "name": "typicode" }
 }
