+++ frontend/src/components/Article/index.ts
@@ -1,0 +1,3 @@
+import Article from "./Article";
+
+export default Article;
